Dodano metodę Remove do ConsoleApp2
A new method named `Remove` was added to the `ConsoleApp2` namespace in `Program.cs`. This method takes a string `s` as a parameter and checks if the last character of the string is an exclamation mark ('!'). If it is, the method returns the string without the last two characters; otherwise, it returns the string unchanged.